Code organisation initial guidelines (#6)
This is an initial attempt to introduce some basic terminology and guidelines for code organisation.

Currently most of the logic is distributed piecemeal throughout forms, tasks, and various other places. It's often duplicated, and frequently inconsistent as parts of the code that have been duplicated have not been updated in lockstep.

I've started here with three main sections:

- Terminology, which tries to concisely define a few DDD terms so that we can use them as common vernacular for talking about code organisation
- Layering, which outlines a layering structure that's typically used for large applications, adapted to Django's idioms
- Guidelines, which currently only contains a single guideline to encapsulate domain logic rather than copy/pasting it, no matter how trivial that domain logic might seem.

There's a lot more I could write here, but I'm hoping to start with something that I think is _relatively_ uncontroversial, and then we can build on that.